requirements.txt
streamlit>=1.28.0
langgraph>=0.0.20
dashscope>=1.14.0
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
typing-extensions>=4.8.0

